feat: add custom `WARC` indexing configuration file
# Motivation

This `config3.conf` specifies how `WARC`s are indexed in `solr`. As of
now, the most important settings to change is the order of language
profiles. Depending on the language of the `WARC` content, the time used
for language detection might take a long time if the matching language
profile is far down on that list.

# Language profile order

The language profiles have been ordered in this sequence:

01. Norwegian(`no`)
02. English(`en`)
03. Estonian(`et`)
04. Swedish(`sv`)
05. Danish(`da`)
06. German(`de`)
07. French(`fr`)
08. Polish(`pl`)
09. Lithuanian(`lt`)
10. Livonian(`lv`)
11. Arabic(`ar`)
12. Turkish(`tr`)
13. Italian(`it`)
14. Spanish(`es`)
15. Russian(`ru`)
16. Ukrainian(`uk`)

# Future work

Note that since `solrwayback` does not have any language profiles for
Northern Sámi(`sme`), Lule Sámi(`smj`), and Southern Sámi(`sma`) readily
available, these languages seem anecdotally misclassified as
Estonian(`et`). Which is why `Estonian` is rather high up on the list.

This could be solved by somehow obtaining the missing language profiles,
but as of now, I do not know how to do so.