feat: index `nettaviser` collection
This commit also added a `nameSuffix` to distinguish the `nettaviser`
job from the previous indexing job.

Since the disk the relevant data is stored is located on a `Persistent
Volume Claim`, called `nettarkivet-4` instead of `nettarkivet-3` we need
to configure this as part of the lowest level overlay.

Note: `nettarkivet-3` represents the files stored on `/nfs/fs03` and is
by default mounted at `/nfs/fs04`. `nettarkivet-4` on the other hand
represents the files stored on `/nfs/fs04` and is mounted at `/nfs/fs04`
by default.