feat: build and release docker images
# Motivation

To easily deploy a `solrwayback` instance in a `kubernetes` cluster, we
need to build at least 3 container images:
`solr`, `solrwayback` and `warc-indexer`.

In short, `solrwayback` is released as a bundle, where the various
configuration of `solr`, `solrwayback` and `warc-indexer` are included.
The various parts of the bundle will be used to build the container
images, which then can be used in a cluster.

The `solr` image will run the underlying index, the `solrwayback` image
will be used for the top level application and the `warc-indexer` image
will be used to index the `warc` files.

# Implementation

## solr

The `solr` image is based on the official `solr` image, but with some
additional configuration taken from the bundle.

## solrwayback

The `solrwayback` image is based on the official `tomcat` image (since
the bundle uses it under the hood), but with some additional
configuration of `CATALINA` taken from the bundle. `CATALINA` is the
main component of `tomcat`.

## warc-indexer

The `warc-indexer` image is based on the official `eclipse-temurin`
since the warc indexing scripts uses `java` under the hood.

## Release

The images are pushed using a `github action` to the `github` container
registry.

# Updating the bundle

To update the bundle one needs to change all instances of top level
`dockerfile` arguments such as
```dockerfile
ARG SOLRWAYBACK_VERSION=4.4.2
ARG SOLRWAYBACK_TOMCAT_VERSION=8.5.60
ARG TOMCAT_TAG=8.5-jdk8-temurin-jammy
```
and in the `release.yml` workflow file to match the new bundle.

```
SOLRWAYBACK_VERSION=4.4.2
SOLR_VERSION=7.7.3
SOLRWAYBACK_TOMCAT_VERSION=8.5.60
TOMCAT_TAG=8.5-jdk8-temurin-jammy
ECLIPSE_TEMURIN_TAG=8-jre
```

Since this is a manual process, it is prone to errors. See the `Future
work` section for a possible solution.

The reason why we do not simply remove the default values of `ARG` in
the `dockerfile` is that it will break the local build. A developer
would need to find the correct build arguments from the `release.yml`
and use them when building locally.

# Future work

## Missing dependencies

The `solrwayback` image currently lack a few dependencies, such as
`ffmpeg`, `imagemagick`, `tesseract-ocr`, `chromium-browser`, etc.

## Configuration sync

The arguments in the `dockerfile`s and the `release.yml` workflow must
be kept in sync, otherwise confusion will arise. To solve this, a
workflow job should be created that reads the `dockerfile`s and the
`release.yml` and ensures that the arguments are in sync.