feat: ensure arguments in sync
# Motivation

Since commit 25360d97b367eff0c75a81ee456f446469892907 we have had to
keep the sets of arguments in sync between the `Dockerfile`s and the
`workflow` files. It is easy to forget to keep the arguments in sync, so
instead this should be verified automatically.

# Implementation

The main idea is to populate a `dict` with the following structure:
```
{
    argumentA: {
        versionA: [PathA, PathB, ...]
        versionB: [PathB, PathC, ...]
        ...
    },
    argumentB: {
        versionX: [PathA, PathB, ...]
        versionY: [PathB, PathC, ...]
        ...
    },
    ...
}
```

After populating the `dict` we must ensure that every argument only have
one version, otherwise we raise an exception.

# Testing

A simple `pytest` test is added to ensure we have basic testing.
This test is also run in a workflow.