ci: add test workflow
# Motivation

Before merging to the default branch, we would want to ensure that the
`dockerfile`s builds successfully.

# Drawbacks

Ideally, this workflow should be identical to
`.github/workflows/release.yml` except for the 'push: True' field and
the triggers. This means that a lot of code is duplicated, which is not
great.

However, this might not be a huge issue as new images are not likely to
be added to this repository.

# Future work

This commit also relies on the `Dockerfile`s and `workflow`s arguments
being kept in sync. This will be improved in a future commit.