feat: Solrwayback 5.x
This commit upgrades solrwayback to version 5.1.0.

See the release notes
https://github.com/netarchivesuite/solrwayback/releases/tag/5.1.0

The purpose of this repository is to build the container images
used to run solrwayback in a containerized environment. In the
name of keeping it simple this commit also removes all uneccessary
files and abstractions that is not needed for this purpose.