feat!: remove chromium-browser from solrwayback image
This commit removes the installation of chromium-browser  from the solrwayback image
because it does not work.

Snap is required to install chromium-browser properly but snap does not play
nice with container builds because systemd is required to run it.

There exists PPA repositories with chromium-browser builds that does not require snap
so it is possibly to get this to work in theory. It will be future work.